fix: Procedure Mapper Model -> Dto 변환시 int Type 캐스팅 제거

+++ sesac7_flutter_recipe_app/lib/data/mapper/procedure_mapper.dart
@@ -14,8 +14,8 @@ extension ProcedureDtoToModel on ProcedureDto {
 extension ProcedureModelToDto on Procedure {
   ProcedureDto toDto() {
     return ProcedureDto(
-      recipeId: recipeId.toInt(),
-      step: step.toInt(),
+      recipeId: recipeId,
+      step: step,
       content: content,
     );
   }
